feat: redesign navbar curtain with proximity grid, rail shapes, and magnetic links
Replace the old curtain (CurvedBottom, ProximityShape floaters, plain MenuItems)
with a new interactive system:

- Proximity grid: invisible 100px grid lines glow near cursor on desktop,
  static SVG grid on mobile/tablet
- Rail system: geometric shapes (squares, circles, triangles) orbit rectangular
  and circular tracks with hover/drag interaction from About section
- Shapes snap to grid intersections on every animation frame
- Magnetic nav links with parallax depth effect (MagneticButton pattern)
- Responsive rail positioning via percentage blueprints snapped to grid
- Theme-aware colors for light/dark mode curtain links
- Remove concept-curtain page (prototyping complete)
